Refactorisation vers une architecture basée sur interfaces
Transition vers une architecture modulaire et testable :
- Refactorisation des tests (`ExchangeFactoryTests.cs`, `CcxtWrapperTests.cs`) pour utiliser `IExchangeWrapper` et des `Mock`.
- Ajout de l'adaptateur `ExchangeAdapter` pour encapsuler les interactions avec `Exchange`.
- Création de l'interface `IExchangeWrapper` pour abstraire les échanges.
- Refactorisation des classes (`ExchangeFactory.cs`, `CcxtWrapper.cs`, `Form1.cs`, `Program.cs`) pour intégrer `IExchangeWrapper`.
- Suppression et réécriture de `IExchangeFactory` dans `Implementation.Utils.Interfaces`.
- Déplacement de `ICcxtWrapper` dans `Implementation.Wrappers.Interfaces`.
- Amélioration de la testabilité et de la modularité du code.

+++ WinFormsApp/Form1.cs
@@ -1,4 +1,4 @@
-using Implementation.Wrappers;
+using Implementation.Wrappers.Interfaces;
 using Microsoft.Extensions.Logging;
 
 namespace WinFormsApp
@@ -28,7 +28,7 @@ namespace WinFormsApp
 
             try
             {
-                decimal usdt = await _ccxt.GetBalanceAsync("USDT");
+                double usdt = await _ccxt.GetBalanceAsync("USDT");
                 label1.Text = $"{usdt:N4} USDT";
                 _logger.LogInformation("Solde USDT récupéré : {Balance}", usdt);
             }
